Split common logic into AbstractRunner.
